nnn 1.2
Changing the test because curses in the DUMB term seems unable to start on new
lines at certain places *and* seems to truncate at 80 columns. Anyway, the gist
is, we were testing for a string that is beyond column 80 and hence is
truncated. Instead we now test for something that's at the beginning of a line.

Closes #15095.

Signed-off-by: ilovezfs <ilovezfs@icloud.com>